LayoutTests/imported/w3c/web-platform-tests/preload/prefetch-headers-expected.txt


FAIL Prefetch should include Sec-Purpose=prefetch and Sec-Fetch-Dest=empty headers assert_equals: expected (string) "prefetch" but got (undefined) undefined
FAIL Prefetch should not include proprietary headers (X-moz/Purpose) assert_false: expected false got true
PASS Prefetch should respect CORS mode
